HSEARCH-3273 More clearly separate tests of the field definition and type definition DSLs
